refactor: use type export, setup testdata
