+++ pkg/timeline/service/fetchMember.test.ts
@@ -1,10 +1,8 @@
 import { Result } from '@mikuroxina/mini-fn';
 import { describe, expect, it, vi } from 'vitest';
 import type { AccountID } from '../../accounts/model/account.js';
-import {
-  Account,
-  dummyAccountModuleFacade,
-} from '../../intermodule/account.js';
+import { dummyAccounts } from '../../accounts/testData/testData.js';
+import { dummyAccountModuleFacade } from '../../intermodule/account.js';
 import { InMemoryListRepository } from '../adaptor/repository/dummy.js';
 import { List, type ListID } from '../model/list.js';
 import { FetchListMemberService } from './fetchMember.js';
@@ -19,32 +17,6 @@ describe('FetchListMemberService', () => {
     createdAt: new Date('2024-01-01T00:00:00.000Z'),
   });
   const repository = new InMemoryListRepository([dummyListData]);
-  const dummyAccountData = [
-    Account.new({
-      id: '11' as AccountID,
-      name: '@johndoe@example.com',
-      nickname: 'John Doe',
-      bio: '',
-      mail: '',
-      role: 'normal',
-      status: 'active',
-      silenced: 'normal',
-      frozen: 'normal',
-      createdAt: new Date('2023-09-10T00:00:00.000Z'),
-    }),
-    Account.new({
-      id: '12' as AccountID,
-      name: '@test@example.com',
-      nickname: 'Test User',
-      bio: '',
-      mail: '',
-      role: 'normal',
-      status: 'active',
-      silenced: 'normal',
-      frozen: 'normal',
-      createdAt: new Date('2023-09-11T00:00:00.000Z'),
-    }),
-  ];
   const service = new FetchListMemberService(
     repository,
     dummyAccountModuleFacade,
@@ -53,13 +25,13 @@ describe('FetchListMemberService', () => {
   it('should fetch list members', async () => {
     vi.spyOn(dummyAccountModuleFacade, 'fetchAccounts').mockImplementation(
       async () => {
-        return Result.ok(dummyAccountData);
+        return Result.ok(dummyAccounts);
       },
     );
 
     const res = await service.handle('1' as ListID);
 
     expect(Result.isOk(res)).toBe(true);
-    expect(Result.unwrap(res)).toStrictEqual(dummyAccountData);
+    expect(Result.unwrap(res)).toStrictEqual(dummyAccounts);
   });
 });
