+++ pkg/timeline/testData/testData.ts
@@ -3,8 +3,7 @@
  * */
 import { Option } from '@mikuroxina/mini-fn';
 
-import { Account, type AccountID } from '../../accounts/model/account.js';
-import type { PartialAccount } from '../../intermodule/account.js';
+import type { AccountID } from '../../accounts/model/account.js';
 import { Note, type NoteID } from '../../notes/model/note.js';
 
 export const dummyPublicNote = Note.new({
@@ -51,22 +50,3 @@ export const dummyDirectNote = Note.new({
   visibility: 'DIRECT',
   createdAt: new Date('2023/10/10 00:00:00'),
 });
-export const dummyAccount1 = Account.new({
-  id: '101' as AccountID,
-  bio: 'this is test user',
-  mail: 'john@example.com',
-  name: '@john@example.com',
-  nickname: 'John Doe',
-  passphraseHash: '',
-  role: 'normal',
-  silenced: 'normal',
-  status: 'active',
-  frozen: 'normal',
-  createdAt: new Date(),
-});
-export const partialAccount1: PartialAccount = {
-  id: dummyAccount1.getID(),
-  name: dummyAccount1.getName(),
-  nickname: dummyAccount1.getNickname(),
-  bio: dummyAccount1.getBio(),
-};
